1.1.0 Finished writing tests. Minor bug fixes. Total coverage in 98% now (excluding domain and DTO layers).

+++ src/main/java/org/retal/domain/OrderRouteProgression.java
@@ -71,7 +71,7 @@ public class OrderRouteProgression {
 
   @Override
   public int hashCode() {
-    return (routeCounter + 1) * (int) Math.pow(id + 1, 2);
+    return (routeCounter + 1) * (int) Math.pow((double) id + 1, 2);
   }
 
   @Override
